完成B和M部分逻辑代码的编写

+++ user/src/main/java/com/cross2u/user/model/base/BaseStorebill.java
@@ -33,11 +33,11 @@ public abstract class BaseStorebill<M extends BaseStorebill<M>> extends Model<M>
 		return get("sbInfo");
 	}
 
-	public void setSbMoney(Long sbMoney) {
+	public void setSbMoney(Integer sbMoney) {
 		set("sbMoney", sbMoney);
 	}
 
-	public Long getSbMoney() {
+	public Integer getSbMoney() {
 		return get("sbMoney");
 	}
